class "TemperatureControl" added. now temperature tracking is entrusted to a specific device + fixed some variables

+++ TrainDR/src/main/java/train/devices/SpeedControl.java
@@ -3,7 +3,7 @@ package train.devices;
 public class SpeedControl {
     double trainSpeed;
 
-    public SpeedControl(int startingSpeed){
+    public SpeedControl(double startingSpeed){
         this.trainSpeed = startingSpeed;
     }
